Reworked Unix connector's internal and fixed tools using it.
